+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/missingoverride/InputMissingOverrideBadAnnotation.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.annotation;
+package com.puppycrawl.tools.checkstyle.checks.annotation.missingoverride;
 
-public class InputBadAnnotationOverride
+public class InputMissingOverrideBadAnnotation
 {
     Runnable r = new Runnable() {
